requirements.txt
beautifulsoup4
requests
aiohttp
openai==0.28
pandas
openpyxl
python-dotenv
shopifyapi

pillow